# api/openapi/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Keystone Core REST API
  description: |
    REST API for Keystone Core - a runtime infrastructure control plane.

    This specification covers all HTTP REST endpoints exposed by Keystone Core components.
    The primary control plane operations use gRPC (see proto files), but HTTP REST endpoints
    are available for cluster management, health checks, file operations, and more.

    ## Authentication

    Keystone Core supports multiple authentication methods:
    - **API Key**: Bearer token in `Authorization` header
    - **mTLS**: Client certificate authentication
    - **JWT**: JSON Web Token authentication

    ## Rate Limiting

    API endpoints are subject to rate limiting. When rate limited, the API returns
    HTTP 429 with `Retry-After` header indicating when to retry.

    ## Versioning

    All REST APIs use the `/api/v1/` prefix. Breaking changes will increment the version.
  version: 0.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  contact:
    name: Keystone Core Team
    url: https://github.com/shawnbutts/keystone-core
servers:
  - url: http://localhost:9090
    description: Local development server
  - url: https://{control-plane-host}:9443
    description: Production server (HTTPS)
    variables:
      control-plane-host:
        default: control-plane.example.com
        description: Control plane hostname
...
security:
  - bearerAuth: []
  - mtls: []
paths:
...
  /api/status:
    get:
      tags:
        - Health
      summary: Server status
      description: |
        Returns detailed server status including version, uptime,
        memory usage, goroutine count, and connected agents.
      operationId: serverStatus
      responses:
        '200':
          description: Server status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # Cluster Management (handler not yet wired into kscore-server)
  # ============================================
...
  /api/v1/agents:
    get:
      tags:
        - Agents
      summary: List agents
      description: |
        Returns a list of all registered agents with their current status.
        Supports filtering by status and labels, and sorting.
      operationId: listAgents
      parameters:
        - name: status
          in: query
          description: Filter by agent status
          schema:
            type: string
            enum: [connected, disconnected, unknown]
        - name: labels
          in: query
          description: Filter by labels (key=value format)
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentList'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/agents/{id}:
    get:
      tags:
        - Agents
      summary: Get agent details
      description: Returns detailed information about a specific agent.
      operationId: getAgent
      parameters:
        - name: id
          in: path
          required: true
          description: Agent ID
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # Execution
  # ============================================
...
  /api/v1/api-keys:
    get:
      tags:
        - API Keys
      summary: List API keys
      description: Returns all configured API keys (keys are not returned, only metadata).
      operationId: listAPIKeys
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIKeyEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - API Keys
      summary: Create API key
      description: Creates a new API key. The key value is only returned once.
      operationId: createAPIKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPIKeyRequest'
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAPIKeyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key authentication
    mtls:
      type: mutualTLS
      description: Mutual TLS client certificate authentication
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
...
    ServerStatus:
      type: object
      properties:
        version:
          type: string
        buildDate:
          type: string
          format: date-time
        goVersion:
          type: string
        uptime:
          type: string
        memoryUsage:
          type: integer
          format: int64
          description: Memory usage in bytes
        goroutines:
          type: integer
        connectedAgents:
          type: integer
...
    Agent:
      type: object
      properties:
        id:
          type: string
        hostname:
          type: string
        os:
          type: string
        arch:
          type: string
        agent_version:
          type: string
        platform_version:
          type: string
        status:
          type: string
          enum: [connected, disconnected, unknown]
        labels:
          type: object
          additionalProperties:
            type: string
        ip_addresses:
          type: array
          items:
            type: string
        ipv4_addresses:
          type: array
          items:
            type: string
        ipv6_addresses:
          type: array
          items:
            type: string
        is_dual_stack:
          type: boolean
        registered_at:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
    AgentList:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        total:
          type: integer
        online:
          type: integer
        offline:
          type: integer
        retrieved_at:
          type: string
          format: date-time
...
    CreateAPIKeyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, operator, readonly]
          default: readonly
        expires_in:
          type: string
          description: Expiration duration (e.g., "30d", "1y", "365d")
          default: "365d"
    CreateAPIKeyResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        key:
          type: string
          description: API key value (only returned once at creation)
        role:
          type: string
        expires_at:
          type: string
          format: date-time
    APIKeyEntry:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        last_used:
          type: string
          format: date-time
...